Fix immediate task execution when defining test cases
I.e. testTask and testCaseTask would start the task as soon as the test definition is created instead of when the test suite is run.
This fixes a breaking change from how testTask used to work and a flaw of the new testCaseTask api.
It is a breaking change to the testCaseTask api.